Create activitymodule package in di package and move it ActivityBuilder and ActivityModule
