+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/AppComponent.kt
@@ -4,6 +4,15 @@ import android.app.Application
 import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjectionModule
+import io.github.droidkaigi.confsched2018.di.activitymodule.AboutThisAppActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.ContributorActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.MainActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.MapActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.SessionDetailActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.SettingsActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.SpeakerDetailActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.SponsorsActivityBuilder
+import io.github.droidkaigi.confsched2018.di.activitymodule.TopicDetailActivityBuilder
 import io.github.droidkaigi.confsched2018.presentation.App
 import javax.inject.Singleton
