+++ app/src/main/java/io/github/droidkaigi/confsched2018/di/activitymodule/MapActivityBuilder.kt
@@ -1,7 +1,8 @@
-package io.github.droidkaigi.confsched2018.di
+package io.github.droidkaigi.confsched2018.di.activitymodule
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
+import io.github.droidkaigi.confsched2018.di.FragmentBuildersModule
 import io.github.droidkaigi.confsched2018.presentation.map.MapActivity
 
 @Module interface MapActivityBuilder {
